Fix the List S (10 cp) display
It messed up the formatting because the credit points were matched as
grades.  Then you had to wrap the status parsing into a condition
because otherwise the (unconditional) matching of the credit points
would read the wrong thing.

In essence the lack of a status in the list messed it up.